+++ src/paco/models/schemas.py
@@ -561,6 +561,7 @@ def IsValidASGAvailabilityZone(value):
         raise InvalidASGAvailabilityZone
     return True
 
+
 # Lambda Environment variables
 class InvalidLambdaEnvironmentVariable(schema.ValidationError):
     __doc__ = 'Can not be a reserved Environment Variable name and must be alphanumeric or _ character only.'
@@ -7454,6 +7455,12 @@ DBParameterGroup
         required=True
     )
 
+class IDBClusterParameterGroup(IDBParameterGroup):
+    """
+DBCluster Parameter Group
+    """
+    pass
+
 class IRDSOptionConfiguration(Interface):
     """
 Option groups enable and configure features that are specific to a particular DB engine.
@@ -7483,7 +7490,7 @@ Option groups enable and configure features that are specific to a particular DB
 
 class IRDS(IResource, IMonitorable):
     """
-RDS Common Interface
+RDS common fields shared by both DBInstance and DBCluster
     """
     @invariant
     def password_or_snapshot_or_secret(obj):
@@ -7500,14 +7507,13 @@ RDS Common Interface
             raise Invalid("Must set one of db_snapshot_identifier, master_user_password or secrets_password for RDS.")
         return True
 
-    allow_major_version_upgrade = schema.Bool(
-        title="Allow major version upgrades",
-        required=False,
-    )
-    auto_minor_version_upgrade = schema.Bool(
-        title="Automatic minor version upgrades",
-        required=False,
-    )
+    @invariant
+    def valid_engine_version(obj):
+        "Validate engine version is supported for the engine type"
+        if obj.engine_version not in gen_vocabulary.rds_engine_versions[obj.engine]:
+            raise Invalid("Engine Version is not support by AWS RDS")
+        return True
+
     backup_preferred_window = schema.TextLine(
         title="Backup Preferred Window",
         required=False,
@@ -7522,11 +7528,7 @@ RDS Common Interface
             title="CloudWatch Log Export",
         ),
         required=False,
-        # ToDo: Constraint that depends upon the database type, not applicable for Aurora
-    )
-    db_instance_type = schema.TextLine(
-        title="RDS Instance Type",
-        required=False,
+        # ToDo: Constraint that depends upon the database type
     )
     db_snapshot_identifier = schema.TextLine(
         title="DB Snapshot Identifier to restore from",
@@ -7542,10 +7544,6 @@ RDS Common Interface
         value_type=schema.Object(IDNS),
         required=False
     )
-    engine = schema.TextLine(
-        title="RDS Engine",
-        required=False,
-    )
     engine_version = schema.TextLine(
         title="RDS Engine Version",
         required=False,
@@ -7554,10 +7552,6 @@ RDS Common Interface
         title="Enable Storage Encryption",
         required=False,
         schema_constraint='Interface'
-    )
-    license_model = schema.TextLine(
-        title="License Model",
-        required=False,
     )
     maintenance_preferred_window = schema.TextLine(
         title="Maintenance Preferred Window",
@@ -7570,16 +7564,6 @@ RDS Common Interface
     master_user_password = schema.TextLine(
         title="Master User Password",
         required=False,
-    )
-    option_configurations = schema.List(
-        title="Option Configurations",
-        value_type=schema.Object(IRDSOptionConfiguration),
-        required=False,
-    )
-    parameter_group = PacoReference(
-        title="RDS Parameter Group",
-        required=False,
-        schema_constraint='IDBParameterGroup'
     )
     primary_domain_name = PacoReference(
         title="Primary Domain Name",
@@ -7594,10 +7578,6 @@ RDS Common Interface
     )
     port = schema.Int(
         title="DB Port",
-        required=False,
-    )
-    publically_accessible = schema.Bool(
-        title="Assign a Public IP address",
         required=False,
     )
     secrets_password = PacoReference(
@@ -7617,6 +7597,41 @@ RDS Common Interface
         required=False,
         schema_constraint='ISegment'
     )
+
+class IRDSInstance(IRDS):
+    """
+RDS DB Instance
+    """
+    allow_major_version_upgrade = schema.Bool(
+        title="Allow major version upgrades",
+        required=False,
+    )
+    auto_minor_version_upgrade = schema.Bool(
+        title="Automatic minor version upgrades",
+        required=False,
+    )
+    db_instance_type = schema.TextLine(
+        title="RDS Instance Type",
+        required=False,
+    )
+    license_model = schema.TextLine(
+        title="License Model",
+        required=False,
+    )
+    option_configurations = schema.List(
+        title="Option Configurations",
+        value_type=schema.Object(IRDSOptionConfiguration),
+        required=False,
+    )
+    parameter_group = PacoReference(
+        title="RDS Parameter Group",
+        required=False,
+        schema_constraint='IDBParameterGroup'
+    )
+    publically_accessible = schema.Bool(
+        title="Assign a Public IP address",
+        required=False,
+    )
     storage_encrypted = schema.Bool(
         title="Enable Storage Encryption",
         required=False,
@@ -7630,8 +7645,7 @@ RDS Common Interface
         required=False,
     )
 
-
-class IRDSMultiAZ(IRDS):
+class IRDSMultiAZ(IRDSInstance):
     """
 RDS with MultiAZ capabilities. When you provision a Multi-AZ DB Instance, Amazon RDS automatically creates a
 primary DB Instance and synchronously replicates the data to a standby instance in a different Availability Zone (AZ).
@@ -7645,20 +7659,106 @@ primary DB Instance and synchronously replicates the data to a standby instance 
 class IRDSMysql(IRDSMultiAZ):
     """RDS for MySQL"""
 
+class IRDSClusterInstance(IParent):
+    """
+DB Instance that belongs to a DB Cluster.
+    """
+    # ToDo: Add an invariant to check that az is within the number in the segment
+    availability_zone = schema.Int(
+        title='Availability Zone where the instance will be provisioned.',
+        description="Must be one of 1, 2, 3 ...",
+        required=False,
+        min=0,
+        max=10,
+    )
+    db_instance_type = schema.TextLine(
+        title="DB Instance Type",
+        required=True,
+    )
+    publicly_accessible = schema.Bool(
+        title="Assign a Public IP address",
+        required=False,
+        default=False,
+    )
+
 class IRDSAurora(IResource, IRDS):
     """
 RDS Aurora
     """
-    secondary_domain_name = PacoReference(
-        title="Secondary Domain Name",
-        str_ok=True,
+    availability_zones = schema.TextLine(
+        title='Availability Zones to launch instances in.',
+        description="Must be one of all, 1, 2, 3 ...",
+        default='all',
         required=False,
-        schema_constraint='IRoute53HostedZone'
+        constraint=IsValidASGAvailabilityZone
     )
-    secondary_hosted_zone = PacoReference(
-        title="Secondary Hosted Zone",
+    backtrack_window_in_seconds = schema.Int(
+        title="Backtrack Window in seconds. Disabled when set to 0.",
+        description="Maximum is 72 hours (259,200 seconds).",
+        min=0,
+        max=259200,
+        default=0,
         required=False,
-        schema_constraint='IRoute53HostedZone'
+    )
+    cluster_parameter_group = PacoReference(
+        title="DB Cluster Parameter Group",
+        required=False,
+        schema_constraint='IDBClusterParameterGroup'
+    )
+    db_instances = schema.List(
+        title="DB Instances",
+        required=False,
+        value_type=schema.Object(IRDSClusterInstance),
+        default=[],
+    )
+    enable_http_endpoint = schema.Bool(
+        title="Enable an HTTP endpoint to provide a connectionless web service API for running SQL queries",
+        default=False,
+        required=False,
+    )
+    engine_mode = schema.Choice(
+        title="Engine Mode",
+        description="Must be one of provisioned, serverless, parallelquery, global, or multimaster.",
+        vocabulary=vocabulary.rds_cluster_engine_mode,
+        required=False,
+    )
+    restore_type = schema.Choice(
+        title="Restore Type",
+        description="Must be one of full-copy or copy-on-write",
+        default='full-copy',
+        vocabulary=vocabulary.rds_restore_types,
+        required=False,
+    )
+    use_latest_restorable_time = schema.Bool(
+        title="Restore the DB cluster to the latest restorable backup time",
+        required=False,
+        default=False,
+    )
+
+class IRDSMysqlAurora(IRDSAurora):
+    """
+RDS MySQL Aurora Cluster
+    """
+    # ToDo: constraint
+    database_name = schema.TextLine(
+        title="Database Name to create in the cluster",
+        description="Must be a valid database name for the DB Engine. Must contain 1 to 64 letters or numbers. Can't be MySQL reserved word.",
+        required=False,
+        min_length=1,
+        max_length=64,
+    )
+
+class IRDSPostgresqlAurora(IRDSAurora):
+    """
+RDS PostgreSQL Aurora Cluster
+    """
+    # ToDo: constraint
+    database_name = schema.TextLine(
+        title="Database Name to create in the cluster",
+        description="Must be a valid database name for the DB Engine. Must contain 1 to 63 letters, numbers or underscores. Must begin with a letter or an underscore. Can't be PostgreSQL reserved word.",
+        required=False,
+        min_length=1,
+        max_length=63,
     )
 
 class IRDSPostgresql(IRDSMultiAZ):